feat: 채팅방 나가기 수정 채팅방 나가기 기능 구현으로 chat 파일들 수정 (#107)

+++ src/main/java/bjs/zangbu/chat/service/ChatServiceImpl.java
@@ -84,13 +84,30 @@ public class ChatServiceImpl implements ChatService{
         }
     }
 
-    //채팅방 삭제
+    //채팅방 나가기
     @Override
     @Transactional
-    public void deleteChatRoom(String chatRoomId) {
-        //chatRoomId의 ChatMessage들 삭제
-        chatMapper.deleteMessagesByRoomId(chatRoomId);
-        //chatRoomId의 ChatRoom 삭제
-        chatMapper.deleteChatRoom(chatRoomId);
+    public void leaveChatRoom(String chatRoomId, String userId) {
+
+        ChatRoom chatRoom = chatMapper.selectChatRoomById(chatRoomId);
+
+        boolean isSeller = userId.equals(chatMapper.selectMemberIdByNickname(userId));
+        boolean isBuyer = userId.equals(chatRoom.getConsumerId());
+
+        if (isSeller) {
+            chatMapper.updateSellerVisible(chatRoomId);
+        } else if (isBuyer) {
+            chatMapper.updateConsumerVisible(chatRoomId);
+        } else {
+            throw new IllegalStateException("채팅방 참여자가 아닙니다.");
+        }
+
+        //일대일 채팅에 참여한 둘 모두 나간 경우에 DB에서 완전 삭제
+        if (!chatRoom.getSeller_visible() && !chatRoom.getConsumer_visible()) {
+            //chatRoomId의 ChatMessage들 삭제
+            chatMapper.deleteMessagesByRoomId(chatRoomId);
+            //chatRoomId의 ChatRoom 삭제
+            chatMapper.deleteChatRoom(chatRoomId);
+        }
     }
 }
